Reorganised the files to UI and Expenses folder
