+++ src/components/UI/Card.tsx
@@ -1,0 +1,7 @@
+import "./Card.css";
+
+function Card(props: any) {
+  return <div className={`card ${props.className}`}>{props.children}</div>;
+}
+
+export default Card;
